+++ src/Conversion.php
@@ -47,7 +47,8 @@ abstract class Conversion
      * Convert Celsius (Cº) To Fahrenheit (Fº).
      *
      * @since  1.2.0
-     * @deprecated current signature and functionality as of 2.1.0, will be changed in 3.0
+     *
+     * @todo current signature and functionality as of 2.1.0, will be changed in 3.0
      *
      * @param float $celsius   Value in Celsius
      * @param bool  $rounded   Whether to round the result.
@@ -64,7 +65,8 @@ abstract class Conversion
      * Convert Celsius (Cº) To Kelvin (K).
      *
      * @since  1.2.0
-     * @deprecated current signature and functionality as of 2.1.0, will be changed in 3.0
+     *
+     * @todo current signature and functionality as of 2.1.0, will be changed in 3.0
      *
      * @param float $celsius   Value in Celsius
      * @param bool  $rounded   Whether to round the result.
@@ -81,7 +83,8 @@ abstract class Conversion
      * Convert Celsius (Cº) To Rankine (ºR).
      *
      * @since  1.2.0
-     * @deprecated current signature and functionality as of 2.1.0, will be changed in 3.0
+     *
+     * @todo current signature and functionality as of 2.1.0, will be changed in 3.0
      *
      * @param float $celsius   Value in Celsius
      * @param bool  $rounded   Whether to round the result.
@@ -98,7 +101,8 @@ abstract class Conversion
      * Convert Fahrenheit (Fº) To Celsius (Cº).
      *
      * @since  1.2.0
-     * @deprecated current signature and functionality as of 2.1.0, will be changed in 3.0
+     *
+     * @todo current signature and functionality as of 2.1.0, will be changed in 3.0
      *
      * @param float $fahrenheit Value in Fahrenheit
      * @param bool  $rounded    Whether to round the result.
@@ -115,7 +119,8 @@ abstract class Conversion
      * Convert Fahrenheit (Fº) To Kelvin (K).
      *
      * @since  1.2.0
-     * @deprecated current signature and functionality as of 2.1.0, will be changed in 3.0
+     *
+     * @todo current signature and functionality as of 2.1.0, will be changed in 3.0
      *
      * @param float $fahrenheit Value in Fahrenheit
      * @param bool  $rounded    Whether to round the result.
@@ -132,7 +137,8 @@ abstract class Conversion
      * Convert Fahrenheit (Fº) To Rankine (ºR).
      *
      * @since  1.2.0
-     * @deprecated current signature and functionality as of 2.1.0, will be changed in 3.0
+     *
+     * @todo current signature and functionality as of 2.1.0, will be changed in 3.0
      *
      * @param float $fahrenheit Value in Fahrenheit
      * @param bool  $rounded    Whether to round the result.
@@ -206,7 +212,8 @@ abstract class Conversion
      * Convert Kelvin (K) To Celsius (Cº).
      *
      * @since  1.2.0
-     * @deprecated current signature and functionality as of 2.1.0, will be changed in 3.0
+     *
+     * @todo current signature and functionality as of 2.1.0, will be changed in 3.0
      *
      * @param float $kelvin    Value in Kelvin
      * @param bool  $rounded   Whether to round the result.
@@ -223,7 +230,8 @@ abstract class Conversion
      * Convert Kelvin (K) To Fahrenheit (Fº).
      *
      * @since  1.2.0
-     * @deprecated current signature and functionality as of 2.1.0, will be changed in 3.0
+     *
+     * @todo current signature and functionality as of 2.1.0, will be changed in 3.0
      *
      * @param float $kelvin    Value in Kelvin
      * @param bool  $rounded   Whether to round the result.
@@ -240,7 +248,8 @@ abstract class Conversion
      * Convert Kelvin (K) To Rankine (ºR).
      *
      * @since  1.2.0
-     * @deprecated current signature and functionality as of 2.1.0, will be changed in 3.0
+     *
+     * @todo current signature and functionality as of 2.1.0, will be changed in 3.0
      *
      * @param float $kelvin    Value in Kelvin
      * @param bool  $rounded   Whether to round the result.
@@ -257,7 +266,8 @@ abstract class Conversion
      * Convert Rankine (ºR) To Celsius (Cº).
      *
      * @since  1.2.0
-     * @deprecated current signature and functionality as of 2.1.0, will be changed in 3.0
+     *
+     * @todo current signature and functionality as of 2.1.0, will be changed in 3.0
      *
      * @param float $rankine   Value in Rankine
      * @param bool  $rounded   Whether to round the result.
@@ -274,7 +284,8 @@ abstract class Conversion
      * Convert Rankine (ºR) To Fahrenheit (Fº).
      *
      * @since  1.2.0
-     * @deprecated current signature and functionality as of 2.1.0, will be changed in 3.0
+     *
+     * @todo current signature and functionality as of 2.1.0, will be changed in 3.0
      *
      * @param float $rankine   Value in Rankine
      * @param bool  $rounded   Whether to round the result.
@@ -291,7 +302,8 @@ abstract class Conversion
      * Convert Rankine (ºR) To Kelvin (K).
      *
      * @since  1.2.0
-     * @deprecated current signature and functionality as of 2.1.0, will be changed in 3.0
+     *
+     * @todo current signature and functionality as of 2.1.0, will be changed in 3.0
      *
      * @param float $rankine   Value in Rankine
      * @param bool  $rounded   Whether to round the result.