+++ tests/src/ArraysTest.php
@@ -13,8 +13,8 @@ declare(strict_types=1);
 
 namespace Esi\Utility\Tests;
 
-use ArrayAccess;
 use Esi\Utility\Arrays;
+use Esi\Utility\Tests\Fixtures\TestArrayAccess;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 use stdClass;
@@ -23,6 +23,8 @@ use stdClass;
  * Array utilities tests.
  *
  * @internal
+ *
+ * @psalm-api
  */
 #[CoversClass(Arrays::class)]
 class ArraysTest extends TestCase
@@ -215,60 +217,5 @@ class ArraysTest extends TestCase
 
         self::assertTrue(Arrays::valueExists($array, 'foo'));
         self::assertFalse(Arrays::valueExists($array, 'bar'));
-    }
-}
-
-/**
- * @implements ArrayAccess<mixed, mixed>
- */
-class TestArrayAccess implements ArrayAccess
-{
-    /**
-     * @var array<int|string, mixed>
-     */
-    public array $container = [
-        'one'   => 1,
-        'two'   => 2,
-        'three' => 3,
-    ];
-
-    /**
-     * Whether an offset exists.
-     */
-    #[\Override]
-    public function offsetExists(mixed $offset): bool
-    {
-        return isset($this->container[$offset]);
-    }
-
-    /**
-     * Retrieve an offset exists.
-     */
-    #[\Override]
-    public function offsetGet(mixed $offset): mixed
-    {
-        return $this->container[$offset] ?? null;
-    }
-
-    /**
-     * Set an offset.
-     */
-    #[\Override]
-    public function offsetSet(mixed $offset, mixed $value): void
-    {
-        if (\is_null($offset)) {
-            $this->container[] = $value;
-        } else {
-            $this->container[$offset] = $value;
-        }
-    }
-
-    /**
-     * Unset an offset.
-     */
-    #[\Override]
-    public function offsetUnset(mixed $offset): void
-    {
-        unset($this->container[$offset]);
     }
 }